// src-srv/protos/service.ts
export interface Block {
    /**
     * ID is the block ID
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * UUID is used to reference another Document in a block.
     *
     * @generated from protobuf field: string uuid = 2;
     */
    uuid: string;
    /**
     * URI is used to reference another entity in a document.
     *
     * @generated from protobuf field: string uri = 3;
     */
    uri: string;
    /**
     * URL is a browseable URL for the the block.
     *
     * @generated from protobuf field: string url = 4;
     */
    url: string;
    /**
     * Type is the type of the block
     *
     * @generated from protobuf field: string type = 5;
     */
    type: string;
    /**
     * Title is the title/headline of the block, typically used in the
     * presentation of the block.
     *
     * @generated from protobuf field: string title = 6;
     */
    title: string;
    /**
     * Data contains block data
     *
     * @generated from protobuf field: map<string, string> data = 7;
     */
    data: {
        [key: string]: string;
    };
    /**
     * Relationship describes the relationship to the document/parent
     * entity
     *
     * @generated from protobuf field: string rel = 8;
     */
    rel: string;
    /**
     * Role is used for nuancing and qualifying relationships.
     *
     * @generated from protobuf field: string role = 9;
     */
    role: string;
    /**
     * Name is a name for the block. An alternative to "rel" when
     * relationship is a term that doesn't fit.
     *
     * @generated from protobuf field: string name = 10;
     */
    name: string;
    /**
     * Value is a value for the block. Useful when we want to store a
     * primitive value.
     *
     * @generated from protobuf field: string value = 11;
     */
    value: string;
    /**
     * ContentType is used to describe the content type of the
     * block/linked entity if it differs from the type of the block.
     *
     * @generated from protobuf field: string contentType = 12;
     */
    contentType: string;
    /**
     * Links are used to link to other resources and documents.
     *
     * @generated from protobuf field: repeated elephant.repository.Block links = 13;
     */
    links: Block[];
    /**
     * Content is used to embed content blocks.
     *
     * @generated from protobuf field: repeated elephant.repository.Block content = 14;
     */
    content: Block[];
    /**
     * Meta is used to embed metadata
     *
     * @generated from protobuf field: repeated elephant.repository.Block meta = 15;
     */
    meta: Block[];
}
// ...
export const Block = new Block$Type();